+++ src/main/java/com/cells/blocks/exportinterface/BlockExportInterface.java
@@ -1,30 +1,6 @@
 package com.cells.blocks.exportinterface;
 
-import java.util.List;
-
-import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
-
-import net.minecraft.block.Block;
-import net.minecraft.block.material.Material;
-import net.minecraft.block.state.IBlockState;
-import net.minecraft.client.resources.I18n;
-import net.minecraft.client.util.ITooltipFlag;
-import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.EnumFacing;
-import net.minecraft.util.EnumHand;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.World;
-
-import net.minecraftforge.fml.relauncher.Side;
-import net.minecraftforge.fml.relauncher.SideOnly;
-
-import appeng.block.AEBaseTileBlock;
-
-import com.cells.Cells;
-import com.cells.Tags;
-import com.cells.core.CellsCreativeTab;
+import com.cells.blocks.interfacebase.AbstractResourceInterfaceBlock;
 import com.cells.gui.CellsGuiHandler;
 
 
@@ -32,45 +8,14 @@ import com.cells.gui.CellsGuiHandler;
  * Block for the Export Interface.
  * A filtered interface that requests items from the network and exposes them for extraction.
  */
-public class BlockExportInterface extends AEBaseTileBlock {
+public class BlockExportInterface extends AbstractResourceInterfaceBlock<TileExportInterface> {
 
     public BlockExportInterface() {
-        super(Material.IRON);
-        this.setRegistryName(Tags.MODID, "export_interface");
-        this.setTranslationKey(Tags.MODID + ".export_interface");
-        this.setCreativeTab(CellsCreativeTab.instance);
-        this.setHardness(2.2F);
-        this.setResistance(6.0F);
-        this.setTileEntity(TileExportInterface.class);
-    }
-
-    @Override
-    @SideOnly(Side.CLIENT)
-    public void addInformation(ItemStack stack, @Nullable World world, List<String> tooltip, ITooltipFlag flag) {
-        super.addInformation(stack, world, tooltip, flag);
-        tooltip.add("§7" + I18n.format("tooltip.cells.export_interface.info"));
-    }
-
-    @Override
-    public boolean onActivated(final World w, final BlockPos pos, final EntityPlayer p,
-                               final EnumHand hand, final @Nullable ItemStack heldItem,
-                               final EnumFacing side, final float hitX, final float hitY, final float hitZ) {
-        if (p.isSneaking()) return false;
-
-        final TileExportInterface tile = this.getTileEntity(w, pos);
-        if (tile != null) {
-            if (!w.isRemote) {
-                p.openGui(Cells.instance, CellsGuiHandler.GUI_EXPORT_INTERFACE, w, pos.getX(), pos.getY(), pos.getZ());
-            }
-            return true;
-        }
-
-        return false;
-    }
-
-    @Override
-    public void neighborChanged(@Nonnull IBlockState state, @Nonnull World worldIn, @Nonnull BlockPos pos,
-                                @Nonnull Block blockIn, @Nonnull BlockPos fromPos) {
-        // Nothing special needed for redstone handling
+        super(
+            "export_interface",
+            "export_interface",
+            TileExportInterface.class,
+            CellsGuiHandler.GUI_EXPORT_INTERFACE
+        );
     }
 }